refactor: 아이콘을 /public/icons 에 넣음에 따라 import 문 변경
#20
